refactor: Extract aws-s3 prefix parameters into a trait to simplify testing
